Feature/eliminate abstract class hierarchy

+++ src/main/java/group/msg/jpowermonitor/dto/Quantity.java
@@ -3,17 +3,15 @@ package group.msg.jpowermonitor.dto;
 import group.msg.jpowermonitor.agent.Unit;
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 /**
  * An amount of a {@link Unit}
  */
 @Value
 public class Quantity {
-    BigDecimal value;
+    Double value;
     Unit unit;
 
-    public static Quantity of(BigDecimal value, Unit unit) {
+    public static Quantity of(Double value, Unit unit) {
         return new Quantity(value, unit);
     }
 }
